history model separate h and cpp
